scanner: Add fallback docs for virtual functions
Many virtual functions definitions are still documented using the
gtk-doc struct field style, as opposed to a full docblock near the
function pointer.

We should still favour the full docblock, as it can contain an
appropriate description and annotations for arguments and return values;
but if there is no block we can at least try to see if there's a field
description to fall back to, to avoid fully undocumented symbols.

Fixes: #478