meson: Fix running when using -Dpkg_config_path on Windows
G-I's build files do not set up PKG_CONFIG_PATH unlike Meson's GNOME
module where it sets up PKG_CONFIG_PATH for us when building
introspection files, when -Dpkg_config_path (or --pkg-config-path) is
being used.

This poses an issue on Windows with Python 3.8.x or later when G-I is built
against GLib where its pkg-config files are not found in pkg-config default
paths or in the PKG_CONFIG_PATH envvar, as We need the help of GLib's
pkg-config files to find the paths where we load the DLLs of GLib and its
dependent DLLs so that the g-ir-* scripts can load them on Python 3.8.x
or later, as Python 3.8.x or later on Windows require DLLs outside of system
paths to have their paths included via using os.add_dll_directory(),
which are done within G-I via envvars or pkg-config files of GLib.

This makes the build files set up for us the PKG_CONFIG_PATH envvar on
Windows if -Dpkg_config_path is being used to locate GLib, so that the
g-ir-* scripts are able to load GLib's DLLs in order to run.

Part of overall fix for issue #492.