Use uninstalled .pc files and %PATH% on Windows as needed
...when GLib is built as a subproject, by setting GLIB_BUILD_DIR so that
the build will look first for the uninstalled gio-2.0-uninstalled.pc
first (if not already so), and then the paths in %PATH% that really
exist.

For builds that build against an existing GLib installation, things will
go on the former way as GLib's dependent DLLs should already be in paths
that are together with the installed GLib DLLs.

This is necessary so that the g-ir-* scripts can locate the
freshly-built GLib and their dependent DLLs so that the scripts can
properly run, as Python 3.8.x or later require non-system DLLs to be in
dedicated paths that are indicated by os.add_dll_directory().