meson: Set GI_EXTRA_BASE_DLL_DIRS if GLib/libffi are subprojects
This way, the g-ir-* scripts can load their DLLs so that things will
run.

Also, double check that proxy-intl and ZLib and PCRE are indeed built as
subprojects when GLib is built a subproject, as they may have been found
in %INCLUDE% and %LIB%, as we can't feed them into
os.add_dll_directory() in this case.  Note that one may still choose to
build the PCRE subproject as DLLs, so we include that as well, and we
want to make sure that these directory names are indeed in-sync with the
settings in GLib's subprojects, so that they do serve their purpose.