dumper: Fix introspection binaries missing rpaths for relative dylibs
Apple expects dylibs to have an ID of the form `@rpath/libyadda.dylib`.
For this convention to work, a RPATH entry pointing to the library
prefix and/or to the relative location of said prefix w.r.t the
dependent executable or library, must be inserted.

This need previously did not arose because gobject-introspection
implicitly relied on the consumers having IDs set to their absolute
paths. This is no longer the case with GStreamer's Cerbero, which now
sanitizes RPATHs, load commands, and dylib IDs at install time.

The simplest and most obvious fix is to ensure that, for each library
load path, the corresponding rpath entry is inserted.