Pair pointer GTypes with their opaque structure
Like we do for boxed types.

We are currently skipping types that derive from G_TYPE_POINTER, mainly
for historical reasonds. GTK4 started using G_TYPE_POINTER types for
opaque structures passed to callbacks and signals by reference to avoid
the cost of a copy/free. This currently has the effect of adding an
untyped data structure to the GIR data, and including the get_type
function in its symbols.

If a G_TYPE_POINTER type has a record or a union associated to it, then
we should pair the GType name and get_type function to the corresponding
structured type in the GIR, instead.

Fixes: #508