Integrate gobject-introspection-tests as git submodule
This takes the approach of a git submodule without a meson subproject,
simply adjusting the paths to the source files so that they point into the
git submodule. Everything else should be the same. We just ignore the
meson.build file in the submodule.

I considered using a meson subproject which would be the recommended way
to integrate g-i-tests for language bindings, but I believe that approach
is fundamentally incompatible with g-i where we need to use the just-built
g-ir-scanner. g-i-tests uses gnome.generate_gir() in its build file, which
uses meson.override_find_program() for g-ir-scanner. That is incompatible
with using it in the parent project to override g-ir-scanner with the
just-built copy.

Closes: #489