Replace _GI_TEST_EXTERN -> GI_TEST_EXTERN and remove gitestmacros.h
When we made the switch to g-i-tests, I took the opportunity to rename
that symbol in the new codebase so that it doesn't start with an
underscore, which is reserved.

Other tests in g-i still use _GI_TEST_EXTERN which is defined in a
separate copy of gitestmacros.h. It's not good to have two copies
floating around, so remove the old copy and use the new symbol. Adjust
include paths to point into g-i-tests where necessary.

Part-of: <https://gitlab.gnome.org/GNOME/gobject-introspection/-/merge_requests/488>