build: Fix Meson deprecation warning
We are generating the GIR file for Cairo, which means we can't use it as
a string when defining the validation test() objects.

To avoid the deprecation warning from Meson, we can switch to an array
of dictionary entries containing a name for the GIR file, and the file
itself, and use the former when generating targets.

Getting rid of the deprecation warning also makes the `meson test`
output a lot nicer, since Meson doesn't have to make up a complicated
test name.