dumper: Add an assertion to help the static analyser
It’s not recommended to run the output of the dumper through a static
analyser (like `scan-build`), but this might sometimes happen as a
result of running an entire code base through it.

Rather than generate some spurious errors, add an assertion to help the
analyser know that if `dump_irepository()` returns false then it will
always set the error.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>