build: Fix introspection with the latest Windows SDK
The latest Windows SDK (10.0.26100.0)+ enforces linking in check symbols
in the binaries that are built using it, which are determined by the CRT
cflag (i.e /MD, /MDd etc) that is used.

As we did not specify a CRT cflag when we run g-ir-scanner, this causes
a problem where compiling and linking the dumper program will fail as
the appropriate check symbols could not be determined.

This attempts to fix this issue by enforcing that a CRT cflag is
specified, based on the value of the 'b_vscrt' option (and therefore,
buildtype, as needed), when we run g-ir-scanner.