giobjectinfo: g_object_info_find_signal(): Avoid repeated allocations
Previously `g_object_info_find_signal()` used `g_object_info_get_signal()`
to retrieve the *i*th signal and compare its name to the desired name.
However, `g_object_info_get_signal()` returns an allocated object.
If the names were not matching, the allocated object was simply dropped,
and this resulted in a lot of unnecessary allocations compared to the
desired number of allocations, which is one.

To avoid much of the overhead pertaining to the creation of these allocated
`GISignalInfo` objects, introduce a new function that inspects the signal
blobs directly and returns an allocated `GISignalInfo` object just for the
matching signal. The function is largely a copy-and-paste of `_g_base_info_find_vfunc()`,
which does the same thing, only for virtual functions.

See https://gitlab.gnome.org/GNOME/glib/-/merge_requests/4382