maintransformer: Warn on transfer-full instance parameter
It's a bit more debatable whether this should be a warning, as many free()
or destroy() methods annotate the instance parameter as transfer-full.
This doesn't apply to memory-managed languages, but it may still apply to
languages with manual memory management. So, use a heuristic to decide
whether to warn or not: don't warn if the method name starts with 'free'
or 'destroy'.