girparser: Silence spurious invalid pointer warning
With a recent version of GCC we get a warning here because the compiler
thinks we may be dereferencing an invalid pointer in `if (*str == '<')` a
couple of lines down, which I guess it thinks is potentially a pointer of
0x0a (null plus strlen("GLib.Error")).

Add a nonnull assertion, which is guaranteed because g_str_has_prefix()
returns FALSE if either of its arguments are null. This silences the
warning.