ast: More special casing for GInitiallyUnowned
Both GInitiallyUnowned and GInitiallyUnownedClass types are aliases to
GObject and GObjectClass, respectively. For historical reasons, we copy
the fields of GObjectClass straight into the GInitiallyUnowned record;
this means that iterating over the fields of GInitiallyUnowned for
virtual functions will yield callbacks that have an instance parameter
with a type that does not match the instance type, which trips the
static virtual function detection.

Since we're special casing GInitiallyUnownedClass, we need to double
down on this particular brand of mess.

Fixes: https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/543