Revert "giscanner: Add support for static virtual functions"
This reverts commit 39101b2728ccebb4eaa8e26f582f6a4d95f22717.

It turns out that we're exposing a bunch of virtual functions from
GObject onto unsuspecting language bindings, like pygobject, and this
ends up breaking applications.

It's way too late in the release cycle to play this kind of games, so
let's revert this functionality for GNOME 48, and re-introduce it early
in GNOME 49, to give time to other projects to adapt.

See: #543