scanner: Prefer some getters over others
At the moment the current set of heuristics to determine a getter for a
property is good for finding *a* getter. However, if there are multiple
candidates we might declare the wrong method as a getter.

We introduce a priority system to determine which getter candidate is
the most appropriate as the getter. The weight were chosen with gaps in
between so that new and better heuristics have space to thrive.

For a property named `p`, these are the possible getter candidates:

 - A method declared via the `(getter p)` annotation
 - The method `get_p`
 - The method `is_p`
 - The method `p`

we declare the getter to be the first candidate in the list for which a
method of the same name is available.

See https://gitlab.gnome.org/GNOME/gjs/-/issues/681.