girepository: Initialize the async functions parameters
When we started supporting the async functions we made the unfortunate
choice of using zero-values as valid parameters for the async and finish
function indexes, and using instead size-of-field-1 (0x3ff) as the unset
value.

This is not a problem for the libraries built with the new
gi-compile-repository, but given that most of libraries are still using
g-ir-compiler for building the typelibs (thanks to meson built-in
support), and that gi-compile-repository is taking long time for being
used everywhere (again, meson is taking some time here), we are building
typelibs that contain invalid data.

This implies that calling gi_callable_info_get_async_function() for a
callable info defined in a typelib built with g-ir-compiler, would just
crash because we will try to fetch an async function at 0 index.

To avoid producing bad typelibs meanwhile, let's then just initialize
these values to the default, not to confuse (or break) girepository-2
consumers.

Not documenting them on purpose, since these are technically not
publicly supported by libgirepository-1